Paramaterized jobfiles with control loop
main.fio.tmpl is the skeleton for the fio jobfile.
control.env is the configuration file. This is normally the only file that would be edited.
run_jobs.sh loops through a list of block sizes and IO types picked up from control.env, and creates job files and pseudo-code "runs" them, putting the file in an output directory.